Added class "Controller24", old code from 2024.
-IF YOU ARE A NEW CODER START HERE-

This is old Code from last 2024's Into the Deep competition.

It was our main code and just reading through it will teach you a lot about how to program a simple LinearOpMode.

Most of the code was written by Kylah, a former member.

This is just a learning resource, please do not edit the code whatsoever.